+++ packages/apps/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * as promises from './promises';
+export * as functionContext from './function-context';