+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @microsoft/teams.ai
 
+## 2.0.0-preview.7
+
+### Patch Changes
+
+- - Add streamable http support to MCPClient
+  - Adds a way to "send" messages from a tab-app to the associated conversation
+- Updated dependencies
+  - @microsoft/teams.common@2.0.0-preview.7
+
 ## 2.0.0-preview.6
 
 ### Patch Changes