+++ tests/a2a/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tests/a2a
 
+## 0.0.1-preview.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @microsoft/teams.a2a@2.0.0-preview.7
+  - @microsoft/teams.ai@2.0.0-preview.7
+  - @microsoft/teams.apps@2.0.0-preview.7
+  - @microsoft/teams.cards@2.0.0-preview.7
+  - @microsoft/teams.dev@2.0.0-preview.7
+  - @microsoft/teams.openai@2.0.0-preview.7
+
 ## 0.0.1-preview.3
 
 ### Patch Changes