+++ tests/tab/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tests/tab
 
+## 0.0.1-preview.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @microsoft/teams.api@2.0.0-preview.7
+  - @microsoft/teams.apps@2.0.0-preview.7
+  - @microsoft/teams.cards@2.0.0-preview.7
+  - @microsoft/teams.client@2.0.0-preview.7
+  - @microsoft/teams.common@2.0.0-preview.7
+  - @microsoft/teams.dev@2.0.0-preview.7
+  - @microsoft/teams.graph@2.0.0-preview.7
+
 ## 0.0.1-preview.6
 
 ### Patch Changes